More work on the item cap
One day, when people read these commit messages, they will understand the many thoughts that went into whether or not this should be brought upon the world.

Since from the day this will be released an unknown form of support hell would break forward, devouring the time of everyone even remotely associated with the CT Developer Role.

Or maybe not.
Probably not.
At least that's the thought that's allowing me to keep moving forward.
I hope that my fears will proof to be unfounded.
If not, ...[the entry ends here]