Prevent access of properties from ZenCode
The reason is that when scripts run, nothing is initialized yet, thus
trying to get properties will lead to issues when resolving the
reference they're bound to. While this works for vanilla, anything
modded may or may not behave correctly: for this reason, it's better to
just prevent access outright. After all, if you want to copy some
properties, it's because you already know their values.

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>